requirements.txt
openai==1.54.4
langchain~=0.3.7
langgraph~=0.2.50
langchain_core
langchain_openai
langchain-community
python-dotenv
aiosqlite
pygments
gradio
streamlit
flask
flask-cors
crewai==0.80.0
crewai_tools
markdown
pdfkit
# pygraphviz
# wkhtmltopdf-pack  # For PDF conversion
tavily-python
